refactor(tests): split research mission coverage
