+++ scripts/run-tests.mjs
@@ -200,7 +200,10 @@ export const SUITES = {
     "src/lib/server/flow-copilot-session.test.ts",
     "src/lib/server/research-mission-store.test.ts",
     "src/lib/server/research-links.test.ts",
-    "src/lib/server/research-mission-runner.test.ts",
+    "src/lib/server/research-mission-runner-lifecycle-actions.test.ts",
+    "src/lib/server/research-mission-runner-reconciliation-evidence.test.ts",
+    "src/lib/server/research-mission-runner-automation-scheduling.test.ts",
+    "src/lib/server/research-mission-runner-concurrency-terminal.test.ts",
     "src/lib/flow/flow-execution-data.test.ts",
     "src/lib/flow/flow-execution-duration.test.ts",
     "src/lib/flow/flow-execution-filters.test.ts",
